PZ-202,PZ-205,PZ-207,PZ-212,PZ-213,PZ-216,PZ-217,PZ-218 | Saving archive should automatically add extension if not already present; Plugin uninstall/purge functionality in Options Screen with Safe Mode option; Language Pack Management; Skinnable UI via CSS; Main screen resize functionality for PearlZip; Extract selected directory; Show Notification Option saves down to the wrong configuration file; Custom bundle base name configuration key should be: configuration.ntak.pearl-zip.custom-resource-bundle
PZ-205:
+ Enabling safe mode only comes into effect after restart and is reflected by title bar

PZ-207:
+ Upon installation of language pack module, it is loaded for use straight away
+ Fixed issue with en-GB language pack and moved to a distinct java package

PZ-212:
+ Loading themes will clear old theme prior to re-installation

PZ-217:
+ Fixed issue with configuration for Show Notifications and where persisted

PZ-218:
+ Fixed configuration issue for custom bundles

PZ-205,PZ-217,PZ-213,PZ-205,PZ-212,PZ-216,PZ-202,PZ-207:
+ TestFX for plugin core cases, JFXSnackBar Notifications, Resize screen, Safe mode, Themes, Language packs, Directory extraction, Save as functionality

+++ pearl-zip-ui/src/main/java/com/ntak/pearlzip/ui/util/ModuleUtil.java
@@ -287,6 +287,22 @@ public class ModuleUtil {
             // Copy themes to local directory
             for (String theme : info.getThemes()) {
                 Path localThemeDir = Paths.get(STORE_ROOT.toAbsolutePath().toString(), "themes");
+                if (Files.exists(localThemeDir.resolve(theme))) {
+                    Files.walkFileTree(localThemeDir.resolve(theme), new SimpleFileVisitor<>() {
+                        @Override
+                        public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+                            Files.deleteIfExists(dir);
+                            return CONTINUE;
+                        }
+
+                        @Override
+                        public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+                            Files.deleteIfExists(file);
+                            return CONTINUE;
+                        }
+                    });
+                }
+                Files.createDirectories(localThemeDir);
                 Files.copy(targetDir.resolve(theme),
                            localThemeDir.resolve(theme),
                            StandardCopyOption.REPLACE_EXISTING);
@@ -448,7 +464,7 @@ public class ModuleUtil {
              .forEach(m -> {
                  try {
                      final Optional<PluginInfo> optPluginInfo = parseManifest(m);
-                     if (names.contains(optPluginInfo.get().getName())) {
+                     if (optPluginInfo.isPresent() && names.contains(optPluginInfo.get().getName())) {
                          Files.deleteIfExists(m);
 
                          synchronized(PLUGINS_METADATA) {
@@ -570,7 +586,7 @@ public class ModuleUtil {
      *  multiple suitable plugins containing Resource Bundle, the bundle is taken in a non-deterministic manner.
      *
      *  @param modulePath
-     *  @param baseName Base name that corresponds with respurce bundle file
+     *  @param baseName Base name that corresponds with resource bundle file
      *  @param locale
      *  @return ResourceBundle - Key-Value pairs holding i18n strings from resource file
      */
